==> code/02_cleaning_sec_mv.ipynb <==
# [Cleaning: SEC dataset, Sector Market Value](#section-title)

#imports
import pandas as pd
from polygon import RESTClient
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings("ignore")

---

## Importing and Consolidationg of past datasets and SEC Dataset:
- The purpose of including the SEC dataset was to obtain a unique identifier (known as the CIK) for each company and reference this across all datasets to ensure I am identifying and matching companies correctly. 

- The GICS dataframe identifies companies by stock ticker and company name, the latter of which can vary by records. For example, "Apple" and "Apple Inc." refer to the same company. To avoid misidentifying identical companies based on name variations, I have amended the GICS dataframe to include each company's CIK identifier, per the __[SEC's database Edgar](https://www.sec.gov/include/ticker.txt)__ .


- CIK, or __[Central Index Key](https://www.sec.gov/edgar/searchedgar/cik)__, is a unique identifier for each exchange-traded stock. Within the context of this project, this key is useful to map companies in the GICS dataframe to this unique code. The CIK is also used by the financial statement API employed below, so encorporatingthe CIK into the GICS dataframe allows for a similar column by which to join the dataframes after calling and cleaning all data. 

# Importing the dataframe for ticker to Cik conversion

df_g = pd.read_csv("../data/cleaned_csvs_interim_steps/gics_cleaned.csv")
df_sec = pd.read_csv("../data/sec_ciks.csv")

df_tickers_list = pd.read_csv("../data/cleaned_csvs_interim_steps/tickers_list.csv")
tickers = df_tickers_list["symbol"].tolist()

tickers

print(df_sec.shape)
df_sec.head()

df_sec = df_sec.rename(columns={'Unnamed: 0': 'ticker', 'Unnamed: 1': 'cik'})

df_sec['cik'].isnull().sum()

df_sec.duplicated().sum()

# Save a clean copy of the df_sec as csv, for record
df_sec.to_csv("../data/cleaned_csvs_interim_steps/sec_cleaned.csv", index = False)

---

## Importing and Consolidationg of Sector Market Value:

# Importing the sector contribution to total market value df:
df_mv = pd.read_csv("../data/gics_sector_chart.csv", index_col = None)

## Deriving Sector Contribution to Total Market Value of US Stock Market

- In order to determine market concentration of each sector and company, I first need to understand the size of the US Financial Markets (in USD) and each sector's contribution to that market.

# Quick Cleaning:
# Converting columns to lowercase, removing whitespace, removing extraneous symbols
df_mv.columns = df_mv.columns.str.replace(' ', '_')
df_mv.columns = df_mv.columns.str.lower()
df_mv.drop(columns = ["url"], inplace = True)
df_mv.head()

df_mv.to_csv('../data/cleaned_csvs_interim_steps/mv_cleaned.csv',index=False)

---

## Concatenating SEC + GICS dataframes, then concatenating (SEC+GICS) + MV

# Convert the data within the common column to uppercase in both dataframes, per convention, to check for matches

df_sec['ticker'] = df_sec['ticker'].str.upper()
df_g["symbol"] = df_g["symbol"].str.upper()

# Using an inner join to merge the dataframes df_sec and df_g to df_sec_g
df_sec_g = pd.merge(df_sec, df_g, left_on='ticker', right_on='symbol')

# Check if all values in the 'ticker' column match the values in the 'symbol' column
all(df_sec_g['ticker'] == df_sec_g['symbol'])

# Check if there are rows where 'ticker' and 'symbol' do not match:

# All rows of 'ticker' and 'symbol' match
df_sec_g.query('ticker != symbol')

### Justification for differences between merged df and df_g count

- Difference in count between merged dataframe and df_g might be explained by companies delisting (buyouts, bankruptcies, etc.) and being updated on the SEC website without being updated on MBA_stocks.com.

1. MBA website error or sec website error
2. lists made at different times, thus recognizing (or not recognizing) bankruptcy, mergers, acquisitions, name changes
3. slightly different "universes" of stocks

df_sec_g.head()

df_sec_g["gics_sector"].unique()

df_sec_g

df_mv["sector"].unique

# Now for the next concatenation:
# Using an inner join to merge the dataframes df_sec and df_g to df_sec_g

df_sec_g_mv = df_sec_g.merge(df_mv, left_on="gics_sector", right_on="sector", how="left")
df_sec_g_mv.drop(columns = ["sector"], inplace = True) # Same as "gics_sector"
df_sec_g_mv.head()

df_sec_g_mv.to_csv('../data/cleaned_csvs_interim_steps/sec_gics_mv_cleaned.csv', index = False)

---
