==> code/03_cleaning_financial_statements.ipynb <==
# [Cleaning: Financial Statements](#section-title)

#imports
import pandas as pd
from polygon import RESTClient
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings("ignore")

---

## Now pulling and cleaning the Financial Statements dataset

The __[Polygon.io Stock Financials vX API](https://polygon.io/docs/stocks/get_vx_reference_financials)__ was used to query the latest company financial statements from all US stock exchanges. Additional documentation is available through __[Polygon's GitHub](https://github.com/polygon-io/client-python)__.

# Must install polygon-api-client before running code
# pip install polygon-api-client

# Bringing the ticker list over from previous notebooks
df_tickers_list = pd.read_csv("../data/cleaned_csvs_interim_steps/tickers_list.csv")
tickers = df_tickers_list["symbol"].tolist()

# My personal API key
polygonAPIkey = "W8sl48pTwN8GCXTVrJ79tIW4tpB7yrDK"

client = RESTClient(polygonAPIkey)

len(tickers)

# Calling the API - Part 1

# empty list
statements = []

# Function to request fin statement data + append to list using list comp

def polygon_api_pull(ticker):
    [statements.append(t) for t in client.vx.list_stock_financials(ticker=ticker, filing_date_gte="2023-05-01")]
    return statements

for ticker in tickers:
    polygon_api_pull(ticker)

# statements
len(statements)

# We use the dir function (= directory) to take a look at the elements within statements
dir(statements[0])

# Calling the API - Part 2

# Assigniing attributes to income statement
attributes_is = [
    "basic_earnings_per_share",
    "cost_of_revenue",
    "gross_profit",
    "operating_expenses",
    "revenues" ]

# Assigniing attributes to statement of comprehensive income
attributes_ci = [
    "comprehensive_income_loss",
    "comprehensive_income_loss_attributable_to_parent",
    "other_comprehensive_income_loss" ]

# Assigning attributes to cash flow statement
attributes_cfs = [
    "exchange_gains_losses",
    "net_cash_flow",
    "net_cash_flow_from_financing_activities" ]

# Assigning asstributes to balance sheet
attributes_bs = [
    "assets",
    "current_assets",
    "liabilities_and_equity",
    "fixed_assets",
    "equity_attributable_to_noncontrolling_interest",
    "noncurrent_assets",
    "liabilities",
    "other_than_fixed_noncurrent_assets",
    "equity_attributable_to_parent",
    "noncurrent_liabilities",
    "current_liabilities",
    "equity" ]

list_fs = []

for i in range(len(statements)): 
    try:
        dict_fs = {}
        dict_fs["cik"] = statements[i].cik
        dict_fs["company_name"] = statements[i].company_name
        dict_fs["fiscal_period"] = statements[i].fiscal_period
        dict_fs["fiscal_year"] = statements[i].fiscal_year
        dict_fs["filing_date"] = statements[i].filing_date


        # Calling attributes of the Income Statement:
        for attr in attributes_is:
            try:
                financials = statements[i].financials.income_statement
                attr_obj = getattr(financials, attr)
                dict_fs["is_" + attr + "_unit"] = attr_obj.unit
                dict_fs["is_" + attr + "_value"] = attr_obj.value
            except Exception as e:
                pass

        # Calling attributes of the Comprehensive Income Statement:
        for attr in attributes_ci:
            try:
                financials = statements[i].financials.comprehensive_income
                attr_obj = getattr(financials, attr)
                dict_fs["ci_" + attr + "_unit"] = attr_obj.unit
                dict_fs["ci_" + attr + "_value"] = attr_obj.value
            except Exception as e:
                pass

        # Calling attributes of the Cash Flow Statement:
        for attr in attributes_cfs:
            try:
                financials = statements[i].financials.cash_flow_statement
                attr_obj = getattr(financials, attr)
                dict_fs["cfs_" + attr + "_unit"] = attr_obj.unit
                dict_fs["cfs_" + attr + "_value"] = attr_obj.value
            except Exception as e:
                pass

        # Calling attributes of the Balance Sheet:
        for attr in attributes_bs:
            try:
                financials = statements[i].financials.balance_sheet.get(attr) # Note that this path is difference bc balance sheet has nested dicts
                dict_fs["bs_" + attr + "_unit"] = financials.unit
                dict_fs["bs_" + attr + "_value"] = financials.value
            except Exception as e:
                pass
        
        list_fs.append(dict_fs)
    except Exception as e:
        pass

df_poly = pd.DataFrame(list_fs)

# Used ChatGPT only to help format the financials variable and call the attribute.
# Referenced Reddit API structure for dictionary and DF creation.

# over 8,000 rows, because the timeline over which we scraped was expanded to ensure that we did not miss companies that appealed for extensions or filed late
#the extended time period over which we scraped means we captured two reports of many companies.
df_poly

# Creating a new column within the data frame = CF - CFF ~ CFO + CFI (as a proxy for CFO)
df_poly["cfs_net_cash_flow_from_op_and_inv"] = df_poly["cfs_net_cash_flow_value"] - df_poly["cfs_net_cash_flow_from_financing_activities_value"]

# Convert dataframe to csv for record and for use in future notebooks:
df_poly.to_csv("../data/cleaned_csvs_interim_steps/polygon_io_output.csv", index=False)

---
