==> code/05_eda_combined_dfs.ipynb <==
# [GICS + SEC + MV Dataset](#section-title)

- This notebook performs the final concatenation of the dataset.

import pandas as pd

import warnings
warnings.filterwarnings("ignore")

# Importing the concatened data frame for SEC, GICS, and market value info
df_sec_g_mv = pd.read_csv("../data/cleaned_csvs_interim_steps/sec_gics_mv_cleaned.csv")

# Importing the financial statement dataframe
df_poly = pd.read_csv("../data/cleaned_csvs_interim_steps/polygon_io_output_cleaned.csv")

## Final Concatenation:(SEC + GICS + MV) + Financial Statements
- The final dataset will also be called "df", as it is the main dataset I will be working on and the simplest name is desirable.

# Rows misalgined as explained in Section 02.
df_sec_g_mv.shape, df_poly.shape

# df_sgmf = df for sec, gics, market values, and financial statements
df = df_sec_g_mv.merge(df_poly, on="cik", how="inner")
df.shape

df.head()

# Convert dataframe to csv for record and for use in future notebooks:
df.to_csv("../data/cleaned_csvs_interim_steps/df_cleaned.csv", index=False)
